test: align dashboard lifecycle and route coverage
